+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Unreleased
 
+#### Improvements
+
+- It is possible to restrict `mix.exs` search (projectDir) to the current root folder only, so in multi-root workspace configurations ElixirLS won't start in the outermost containing the `mix.exs` folder
+
 ### v0.22.1: 10 July 2024
 
 #### Improvements